move some parts of implementation to common
