+++ src/mono/System.Private.CoreLib/src/System/Runtime/InteropServices/GCHandle.Mono.cs
@@ -8,15 +8,15 @@ namespace System.Runtime.InteropServices
     public partial struct GCHandle
     {
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
-        private static extern IntPtr InternalAlloc(object? value, GCHandleType type);
+        internal static extern IntPtr InternalAlloc(object? value, GCHandleType type);
 
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
-        private static extern void InternalFree(IntPtr handle);
+        internal static extern void InternalFree(IntPtr handle);
 
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
-        private static extern object? InternalGet(IntPtr handle);
+        internal static extern object? InternalGet(IntPtr handle);
 
         [MethodImplAttribute(MethodImplOptions.InternalCall)]
-        private static extern void InternalSet(IntPtr handle, object? value);
+        internal static extern void InternalSet(IntPtr handle, object? value);
     }
 }
